Updated project as per the guidelines and used lombok to reduce boilerplate code
