--- src/main/java/com/koteswararaobudati/Student/StudentClass/Student.java
@@ -1,95 +1,0 @@
-package com.koteswararaobudati.Student.StudentClass;
-
-import jakarta.persistence.*;
-import java.time.LocalDate;
-import java.time.Period;
-
-@Entity
-@Table(
-        name = "Student"
-)
-public class Student {
-    @Id
-    @SequenceGenerator(
-            name = "Student_sequence",
-            sequenceName = "Student_sequence",
-            allocationSize = 1
-    )
-    @GeneratedValue(
-            strategy = GenerationType.SEQUENCE,
-            generator = "Student_sequence"
-    )
-    private long id;
-    private String name;
-    private String email;
-    private LocalDate dob;
-    @Transient
-    private int age;
-
-    public Student() {
-    }
-
-    public Student(long id, String name, String email, LocalDate dob)  {
-        this.id = id;
-        this.name = name;
-        this.email = email;
-        this.dob = dob;
-    }
-
-
-
-    public Student(String name, String email, LocalDate dob) {
-        this.name = name;
-        this.email = email;
-        this.dob = dob;
-    }
-    @Override
-    public String toString() {
-        return "Student{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                ", email='" + email + '\'' +
-                ", dob=" + dob +
-                ", age=" + age +
-                '}';
-    }
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public LocalDate getDob() {
-        return dob;
-    }
-
-    public void setDob(LocalDate dob) {
-        this.dob = dob;
-    }
-
-    public int getAge() {
-        return Period.between(this.dob, LocalDate.now()).getYears();
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-}
