+++ src/main/java/com/koteswararaobudati/Student/config/StudentConfig.java
@@ -1,4 +1,4 @@
-package com.koteswararaobudati.Student.StudentClass;
+package com.koteswararaobudati.Student.config;
 
 
 import org.springframework.context.annotation.Configuration;
