+++ src/main/java/com/koteswararaobudati/Student/controller/FacultyController.java
@@ -1,5 +1,7 @@
-package com.koteswararaobudati.Student.Faculty;
+package com.koteswararaobudati.Student.controller;
 
+import com.koteswararaobudati.Student.service.FacultyService;
+import com.koteswararaobudati.Student.entity.Faculty;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -10,12 +12,9 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/Faculty/")
 public class FacultyController {
 
-    private final FacultyService facultyService;
-
     @Autowired
-    public FacultyController(FacultyService facultyService) {
-        this.facultyService = facultyService;
-    }
+    private FacultyService facultyService;
+
 
     @PostMapping("registerFaculty/")
     public void registerFaculty(@RequestBody Faculty faculty){
