+++ src/main/java/com/koteswararaobudati/Student/repository/StudentRepository.java
@@ -1,5 +1,6 @@
-package com.koteswararaobudati.Student.StudentClass;
+package com.koteswararaobudati.Student.repository;
 
+import com.koteswararaobudati.Student.entity.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
